pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lox"
version = "0.1.0"
description = "Logging library for JAX that compatible with transformations such as jit and vmap."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
  { name = "huterguier", email = "huterguier@gmail.com" }
]
dependencies = [
  "jax>=0.6.0",
  "rich",
  "wandb>=0.20.1",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
cuda = ["jax[cuda12]>=0.6.0"]
dev = [
  "black",
  "isort",
  "pytest",
  "pytest-cov",
]
test = [
  "pytest",
  "pytest-cov",
]
docs = [
  "matplotlib",
  "absl-py",
  "ipython>=8.8.0",  # 8.7.0 has ipython3 lexer error
  "pydata-sphinx-theme==0.14.4",  # v0.15 breaks sidebar toggling
  "snowballstemmer<3.0.0",  # v3.0.0 incompatible with older sphinx; missing stemmer
  "sphinx>=7.3.2,<8.0",  # 7.3.0 breaks sphinx-book-theme; 8.0 breaks myst-nb 1.1
  "sphinx-book-theme==1.1.1",  # v1.1.2 requires pydata-sphinx-theme v0.15
  "sphinx-copybutton>=0.5.0",
  "sphinx-remove-toctrees",
  "sphinx-design",
  "sphinxext-rediraffe",
  "myst-nb>=1.0.0",
  "sphinx-autodoc-typehints",
]

[project.urls]
Homepage = "https://github.com/huterguier/lox"

[tool.setuptools.packages.find]
include = ["lox", "lox.*"]
exclude = ["images"]
